api/Naive-Bayes/requirements.txt
# Function dependencies, for example:
# package>=version
google-auth==1.32.1
google-auth-oauthlib==0.4.4
google-pasta==0.2.0
googleapis-common-protos==1.53.0
importlib-resources==5.2.0
Keras-Preprocessing==1.1.2
numpy==1.19.5
pandas==1.3.3
requests==2.26.0
requests-oauthlib==1.3.0
google-cloud-storage==1.42.3
flask
joblib
scikit-learn